clean up webserver.go
